requirements.txt
numpy==1.18.5
tensorflow-gpu==2.3.0
absl-py==0.11.0
astor==0.8.1
astunparse==1.6.3
chardet==3.0.4
h5py==2.10.0
importlib-metadata==2.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib==3.3.2
pandas==1.1.4
Pillow==8.0.1
pydot==1.4.1
tensorflow-estimator==2.3.0
six==1.15.0